+++ app/sidekiq/refresh_free_tier_credits_job.rb
@@ -33,7 +33,6 @@ class RefreshFreeTierCreditsJob
   REFRESHABLE_PLAN_TYPES = %w[
     free
     basic_trial
-    myspeak
     basic
     pro
     partner_pro